Split main script into separate modules
