+++ client/index.js
@@ -1,0 +1,8 @@
+
+import connect from './connect.js';
+
+const client = {
+    connect
+};
+
+export default client;
