+++ server/index.js
@@ -1,0 +1,10 @@
+
+const ROUTE_ACCEPT_RULES = '/accept-rules';
+const ROUTE_REJECT_RULES = '/reject-rules';
+
+export default {
+    routes: {
+        ROUTE_ACCEPT_RULES,
+        ROUTE_REJECT_RULES
+    }
+};
