+++ server/routes.js
@@ -1,0 +1,8 @@
+
+const ROUTE_ACCEPT_RULES = '/accept-rules';
+const ROUTE_REJECT_RULES = '/reject-rules';
+
+export default {
+    ROUTE_ACCEPT_RULES: ROUTE_ACCEPT_RULES,
+    ROUTE_REJECT_RULES: ROUTE_REJECT_RULES
+};
