+++ validation/generateId.js
@@ -1,0 +1,10 @@
+const alphaNumber = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
+const generateId = (length = 26, chars= alphaNumber) => {
+    let result = '';
+    for (let i = length; i > 0; --i) {
+        result += chars[Math.floor(Math.random() * chars.length)];
+    }
+    return result;
+};
+
+export default generateId;
