+++ workflow/actions.js
@@ -1,0 +1,8 @@
+
+const ACTION_ID_ACCEPT_RULES = 'accept';
+const ACTION_ID_REJECT_RULES = 'reject';
+
+export default {
+    ACTION_ID_ACCEPT_RULES: ACTION_ID_ACCEPT_RULES,
+    ACTION_ID_REJECT_RULES: ACTION_ID_REJECT_RULES
+};
